fix: export more from swap
